Restructured package directory structure.
