+++ build_local.bat
@@ -1,0 +1,6 @@
+rm dist/*
+
+python -m build
+
+python -m twine check dist/*
+
