+++ src/kronoterm_cloud_api/enums/hp_enums.py
